raefactor: form as separate component
