+++ src/components/Form/Form.module.scss
@@ -1,4 +1,3 @@
-
 .form {
   width: 600px;
   .form-input {
